Add MetaStrategy abstract class
