+++ solver/src/main/java/org/chocosolver/solver/search/strategy/strategy/ConflictOrderingSearch.java
@@ -31,21 +31,12 @@ import java.util.*;
  * @author Charles Prud'homme
  * @since 15/06/2016
  */
-public class ConflictOrderingSearch<V extends Variable> extends AbstractStrategy<V> implements IMonitorContradiction {
+public class ConflictOrderingSearch<V extends Variable> extends MetaStrategy<V> implements IMonitorContradiction {
 
     //***********************************************************************************
     // VARIABLES
     //***********************************************************************************
 
-    /**
-     * The target solver
-     */
-    protected Model model;
-
-    /**
-     * The main strategy declared in the solver
-     */
-    private final AbstractStrategy<V> mainStrategy;
     /**
      * Store the variables in conflict
      */
@@ -80,9 +71,7 @@ public class ConflictOrderingSearch<V extends Variable> extends AbstractStrategy
      * @param mainStrategy the main strategy declared
      */
     public ConflictOrderingSearch(Model model, AbstractStrategy<V> mainStrategy) {
-        super(mainStrategy.vars);
-        this.model = model;
-        this.mainStrategy = mainStrategy;
+        super(model, mainStrategy);
         // internal datastructures
         vars = new ArrayList<>();
         var2pos = new TIntIntHashMap(16, .5f, -1, -1);
@@ -90,39 +79,12 @@ public class ConflictOrderingSearch<V extends Variable> extends AbstractStrategy
         next = new TIntArrayList();
         pcft = -1;
         this.scope = new HashSet<>(Arrays.asList(mainStrategy.vars));
+        active = true;
     }
 
-    //***********************************************************************************
-    // METHODS
-    //***********************************************************************************
 
-    @Override
-    public boolean init() {
-        if(model.getSolver().getSearchMonitors().contains(this)) {
-            model.getSolver().plugMonitor(this);
-        }
-        return mainStrategy.init();
-    }
-
-    @Override
-    public void remove() {
-        this.mainStrategy.remove();
-        if(model.getSolver().getSearchMonitors().contains(this)) {
-            model.getSolver().unplugMonitor(this);
-        }
-    }
-
-    @SuppressWarnings("unchecked")
-    @Override
-    public Decision<V> getDecision() {
-        V decVar = firstNotInst();
-        if (decVar != null) {
-            Decision d = mainStrategy.computeDecision(decVar);
-            if (d != null) {
-                return d;
-            }
-        }
-        return mainStrategy.getDecision();
+    public V getSelectedVariable() {
+        return firstNotInst();
     }
 
     //***********************************************************************************
@@ -134,7 +96,7 @@ public class ConflictOrderingSearch<V extends Variable> extends AbstractStrategy
     public void onContradiction(ContradictionException cex) {
         //noinspection unchecked
         Decision<V> dec = model.getSolver().getDecisionPath().getLastDecision();
-        if(dec != RootDecision.ROOT) {
+        if (dec != RootDecision.ROOT) {
             if (scope.contains(dec.getDecisionVariable())) {
                 stampIt(dec.getDecisionVariable());
             }
@@ -191,15 +153,15 @@ public class ConflictOrderingSearch<V extends Variable> extends AbstractStrategy
         return null;
     }
 
-    boolean check(){
+    boolean check() {
         boolean ok = true;
         int first = -1;
-        for(int i = 0; i < vars.size() && ok; i++){
+        for (int i = 0; i < vars.size() && ok; i++) {
             int p = prev.get(i);
             int n = next.get(i);
             ok = (i == pcft && n == -1) || prev.get(n) == i;
             ok &= p == -1 || next.get(p) == i;
-            if(p == -1){
+            if (p == -1) {
                 ok &= first == -1;
                 first = i;
             }
