+++ solver/src/main/java/org/chocosolver/solver/search/strategy/strategy/LastConflict.java
@@ -11,10 +11,8 @@ package org.chocosolver.solver.search.strategy.strategy;
 
 import org.chocosolver.solver.Model;
 import org.chocosolver.solver.exception.ContradictionException;
-import org.chocosolver.solver.search.loop.monitors.IMonitorContradiction;
 import org.chocosolver.solver.search.loop.monitors.IMonitorRestart;
 import org.chocosolver.solver.search.loop.monitors.IMonitorSolution;
-import org.chocosolver.solver.search.strategy.decision.Decision;
 import org.chocosolver.solver.variables.Variable;
 
 import java.util.Arrays;
@@ -29,26 +27,11 @@ import java.util.Set;
  * @author Jean-Guillaume Fages, Charles Prud'homme
  * @since 03/05/2013
  */
-public class LastConflict<V extends Variable> extends AbstractStrategy<V> implements IMonitorRestart, IMonitorSolution, IMonitorContradiction {
+public class LastConflict<V extends Variable> extends MetaStrategy<V> implements IMonitorRestart, IMonitorSolution {
 
     //***********************************************************************************
     // VARIABLES
     //***********************************************************************************
-
-    /**
-     * The target solver
-     */
-    protected Model model;
-
-    /**
-     * The main strategy declared in the solver
-     */
-    private final AbstractStrategy<V> mainStrategy;
-
-    /**
-     * Set to <tt>true</tt> when this strategy is active
-     */
-    protected boolean active;
 
     /**
      * Number of conflicts stored
@@ -68,18 +51,17 @@ public class LastConflict<V extends Variable> extends AbstractStrategy<V> implem
 
     /**
      * Creates a last conflict heuristic
-     * @param model the solver to attach this to
+     *
+     * @param model        the solver to attach this to
      * @param mainStrategy the main strategy declared
-     * @param k the maximum number of conflicts to store
+     * @param k            the maximum number of conflicts to store
      */
     public LastConflict(Model model, AbstractStrategy<V> mainStrategy, int k) {
-        super(mainStrategy.vars);
+        super(model, mainStrategy);
         assert k > 0 : "parameter K of last conflict must be strictly positive!";
-        this.model = model;
-        this.mainStrategy = mainStrategy;
         this.scope = new HashSet<>(Arrays.asList(mainStrategy.vars));
         //noinspection unchecked
-        conflictingVariables = (V[])new Variable[k];
+        conflictingVariables = (V[]) new Variable[k];
         nbCV = 0;
         active = false;
     }
@@ -88,37 +70,10 @@ public class LastConflict<V extends Variable> extends AbstractStrategy<V> implem
     // METHODS
     //***********************************************************************************
 
-    @Override
-    public boolean init(){
-        if(!model.getSolver().getSearchMonitors().contains(this)) {
-            model.getSolver().plugMonitor(this);
-        }
-
-        return mainStrategy.init();
-    }
 
     @Override
-    public void remove() {
-        this.mainStrategy.remove();
-        if(model.getSolver().getSearchMonitors().contains(this)) {
-            model.getSolver().unplugMonitor(this);
-        }
-    }
-
-    @SuppressWarnings("unchecked")
-    @Override
-    public Decision<V> getDecision() {
-        if (active) {
-            V decVar = firstNotInst();
-            if (decVar != null) {
-                Decision<V> d = mainStrategy.computeDecision(decVar);
-                if (d != null) {
-                    return d;
-                }
-            }
-        }
-        active = true;
-        return mainStrategy.getDecision();
+    public V getSelectedVariable() {
+        return firstNotInst();
     }
 
     //***********************************************************************************
@@ -142,9 +97,6 @@ public class LastConflict<V extends Variable> extends AbstractStrategy<V> implem
         }
     }
 
-    @Override
-    public void beforeRestart() {
-    }
 
     @Override
     public void afterRestart() {
